133: Add support for an explicit EXPORTED_SITE_URL in settings
The value of EXPORTED_SITE_URL must match the root URL of the developerportal Site in Wagtail's configuration, because THAT value (Site.hostname) is what determines the domain used in any absolute URLs generated when baking out the site.

(As such, that Site.hostname should not be set to the same value as BASE_URL/where the CMS is actually running, else you end up with the static site's links pointing back to the live CMS.)

OK, but if that URL is something we need to configure in Wagtail, why have it in settings, too?
Because we must add it to ALLOWED_HOSTS, else we will hit `DisallowedHost:Invalid HTTP_HOST header` during baking in production mode – which then outputs all pages saying "400 Bad Request".