133: Add support for an explicit EXPORTED_SITE_URL in settings
The value of EXPORTED_SITE_URL must match the hostname and scheme/port spec of the default developerportal `Site` in Wagtail's settings. That value (Site.hostname) is what determines the domain used for the request when baking out the site, which is then used by `get_scheme_and_host` to generate absolute URLs (eg for the Open Graph tags).

(As such, that Site.hostname should not be set to the same value as BASE_URL/where the CMS is actually running, else you end up with the static site's Open Graph links pointing back to the live CMS.)

The value of EXPORTED_SITE_URL is added to ALLOWED_HOSTS in production settings to avoid hitting `DisallowedHost:Invalid HTTP_HOST header` during baking in production mode – which then outputs all pages saying "400 Bad Request".