133: Update k8s staging config with EXPORTED_SITE_URL
Now we have a new setting, this changeset makes sure we populate it for staging.

IMPORTANT: given that stage.sh mentions `dev` not `stage` (which stage-oregon does mention) I've made up a FQDN for the `dev` deployment, should we choose to set it up.

Having both of them point to the same published/static staging FQDN didn't seem sensible, because of the confusion it would cause.