133: Address problem of static/S3-site URLs linking back to Wagtail live views
This changeset (when combined with a correct configuration for Site.hostname in Wagtail) fixes the issue where the baked site was linking back to live wagtail CMS.

--

Removing the use in templates of a `base_url` value that points to the Wagtail CMS means the URLs go from being absolute to relative, which means they are always appropriate for the version of the site being browsed, whether it's the live Wagtail-hosted one or the S3-hosted baked site.

In addition, for the moments where the hostname is required to build an absolute URL, the main Site.hostname value should be set to the URL of the site where the baked site will be deployed, rather than the CMS URL where it's being made. (Wagtail can still serve live pages on that CMS URL, even when the hostname is set to the S3 site's domain).

Related to the above, there is one area (Open Graph <meta> tags) where absolute URLs are needed. For this, I've just added a simple template tag which uses the HttpRequest to get the scheme and hostname, which works when the site is being baked out for copying to S3.