225: Ensure featured cards have fallback images (#239)
If an Article Page's detail and its Card detail both lack an image, prior to this commit there was no fallback image in use.

This changeset addresses that by using the standard placeholder image when nothing better is available.

Note that the 960px-wide image is used for both the 480w (card-featured.html) and 1024w (card-featured-primary.html) use-cases, because we can't create Wagtail renditions of images that are not known of in the database. I am assuming likelihood of a primary-featured article lacking an image is low, so I think we can make do with 960px being upscaled to 1024px, partic as the 960px may already be cached. If I'm wrong, we can add a 1024px version to `src/img/placeholders/`