133: Address problem of static/S3-site URLs linking back to Wagtail live views (#204)
* 133: Add support for an explicit EXPORTED_SITE_URL in settings

The value of EXPORTED_SITE_URL must match the hostname and scheme/port spec of the default developerportal `Site` in Wagtail's settings. That value (Site.hostname) is what determines the domain used for the request when baking out the site, which is then used by `get_scheme_and_host` to generate absolute URLs (eg for the Open Graph tags).

(As such, that Site.hostname should not be set to the same value as BASE_URL/where the CMS is actually running, else you end up with the static site's Open Graph links pointing back to the live CMS.)

The value of EXPORTED_SITE_URL is added to ALLOWED_HOSTS in production settings to avoid hitting `DisallowedHost:Invalid HTTP_HOST header` during baking in production mode – which then outputs all pages saying "400 Bad Request".

* 133: Update k8s staging config with EXPORTED_SITE_URL

Now we have a new setting, this changeset makes sure we populate it for staging.

IMPORTANT: given that stage.sh mentions `dev` not `stage` (which stage-oregon does mention) I've made up a FQDN for the `dev` deployment, should we choose to set it up.

Having both of them point to the same published/static staging FQDN didn't seem sensible, because of the confusion it would cause.

* 133: Address problem of static/S3-site URLs linking back to Wagtail live views

This changeset (when combined with a correct configuration for Site.hostname in Wagtail) fixes the issue where the baked site was linking back to live wagtail CMS.

--

Removing the use in templates of a `base_url` value that points to the Wagtail CMS means the URLs go from being absolute to relative, which means they are always appropriate for the version of the site being browsed, whether it's the live Wagtail-hosted one or the S3-hosted baked site.

In addition, for the moments where the hostname is required to build an absolute URL, the main Site.hostname value should be set to the URL of the site where the baked site will be deployed, rather than the CMS URL where it's being made. (Wagtail can still serve live pages on that CMS URL, even when the hostname is set to the S3 site's domain).

Related to the above, there is one area (Open Graph <meta> tags) where absolute URLs are needed. For this, I've just added a simple template tag which uses the HttpRequest to get the scheme and hostname, which works when the site is being baked out for copying to S3.