Refactor how we enqueue jobs in Celery, to avoid concurrent builds and race conditions
Prior to this commit, if a flurry of build reuquets were issued (eg multiple authors publishing)
this would put N jobs in the queue. Celery is configured by default to use the number of cores
on its worker node, so this would mean 4 (or more) builds taking place at once, and all targetting
the same output bucket in S3.

IFF one could trust that the order of execution was stable, this would be tolerable (if wasteful)
but because it's async, there's no guarantee that Worker 4 will finish after Workers 1, 2 and 3.
Plus, this would cause four calls to invalidate the entire CDN, which itself is a bit of a b
lack-box experience as it rolls from "InProgress" to "Complete".

So, a new approach is needed.

I've taken a few swings at this, most of which were just not satisfactory. (eg, in-task code that
checks for other running tasks of the same type is still prone to stale state and therefore bad
decisions about whether to terminate itself or carry on, etc). Running celery with --pool=solo
to only do one thing at a time would quickly cause problems as soon as we needed to do _other_
things with a task queue, or spin up multiple worker nodes, etc.

So, instead, the approach has changed to one where builds are requested, and then that signal
is checked frequently and exclusively actioned if it's neeeded.

We do this using a sentinel value, stored in a shared (Redis-backed) cache that, when True,
signifies that a build is required. Separately from this, we have a regular scheduled
task (via celery-beat) that calls the _static_build_async command every minute. That
static-build function has been updated to
a) check for the sentinel to decide whether to do a build or not and to un-set it if it does
go ahead and start a build-and-sync to S3;
b) use a distributed lock to ensure that no other builds can take place while this build is
running;
c) if a build is skipped because one is running, ensure the sentinel is set again so that the
next time the scheduler tries, it will still be clear that a fresh build is needed.

This means that for a flurry of, say, 8 requests, request 1 will trigger a build, while the
rest will just (idempotently) set the sentinel to point out that someone else wanted a build
while the current one was completing.