Fixes certain race conditions on the ci test jobs
There are certain conditions that we run into that causes ci jobs fail
and its generally caused by either the database not being ready or a
permission denied error, which to me indicates that the old credentials
are still lingering after a job completes.

This commit hopes to fix that by using the `urlwait` tool to wait for
the database instance to be up and running as well as additional steps
in there that will go ahead and cleanup the environment before a build
and run.