Support HTTPS for local development.
Adds nginx + SSL cert generation into the local docker stack, following how we do it for kuma https://github.com/mdn/kuma/blob/master/docker-compose.ssl.yml

http://localhost:8000 still responds, but ideally all local development should be done on https://developer-portal-127-0-0-1.nip.io so that future changes (eg CSP support) don't break belocal behaviour

Local fixtures are updated so that the default Wagtail site is expecting to be on `developer-portal-127-0-0-1.nip.io`, and this is matched by BASE_URL, so that the static build still works.

Tried a static build and it still works fine.

Nothing in this changeset should affect non-local behaviour