279: Re-architect how we're using our task queue to avoid concurrent builds (#412)
* Refactor how we enqueue jobs in Celery, to avoid concurrent builds and race conditions

Prior to this commit, if a flurry of build reuquets were issued (eg multiple authors publishing)
this would put N jobs in the queue. Celery is configured by default to use the number of cores
on its worker node, so this would mean 4 (or more) builds taking place at once, and all targetting
the same output bucket in S3.

IFF one could trust that the order of execution was stable, this would be tolerable (if wasteful)
but because it's async, there's no guarantee that Worker 4 will finish after Workers 1, 2 and 3.
Plus, this would cause four calls to invalidate the entire CDN, which itself is a bit of a b
lack-box experience as it rolls from "InProgress" to "Complete".

So, a new approach is needed.

I've taken a few swings at this, most of which were just not satisfactory. (eg, in-task code that
checks for other running tasks of the same type is still prone to stale state and therefore bad
decisions about whether to terminate itself or carry on, etc). Running celery with --pool=solo
to only do one thing at a time would quickly cause problems as soon as we needed to do _other_
things with a task queue, or spin up multiple worker nodes, etc.

So, instead, the approach has changed to one where builds are requested, and then that signal
is checked frequently and exclusively actioned if it's neeeded.

We do this using a sentinel value, stored in a shared (Redis-backed) cache that, when True,
signifies that a build is required. Separately from this, we have a regular scheduled
task (via celery-beat) that calls the _static_build_async command every minute. That
static-build function has been updated to
a) check for the sentinel to decide whether to do a build or not and to un-set it if it does
go ahead and start a build-and-sync to S3;
b) use a distributed lock to ensure that no other builds can take place while this build is
running;
c) if a build is skipped because one is running, ensure the sentinel is set again so that the
next time the scheduler tries, it will still be clear that a fresh build is needed.

This means that for a flurry of, say, 8 requests, request 1 will trigger a build, while the
rest will just (idempotently) set the sentinel to point out that someone else wanted a build
while the current one was completing.

* Amend management command to call a more appropriate build-and-sync command
...so that we don't risk concurrent builds

* Tidying up, including standardising on one piece of locking code

While django-redis's cache.lock() seems legit (it's just wrapping Redis) it makes
sense to use a single locking implementation, and because we need a return value
regarding whether the lock is acquired (which cache.lock) doesn't do while it's
working as a context manager, the redis_lock implemenation based on the Celery
cookbook is the one I picked from the two

* Minor phrasing and labelling fixups to improve logs

* Add test settings based on production ones (not dev) and make the cache safe
+ ensure CI tests use the test settings

* Move some sentinel checking into its own function, for easier testing

* Stop passing the `force` kwarg in from the static_build chain - no longer relevant given the build-request approach

* Add tests for all of the static-build tasks/funcs

The only thing that lacks tests is .context_managers.redis_lock() - for now...

* Add light tests for the redis_lock context manager

* Fixup test regressions in this changeset

* Update k8s config to spin up a celerybeat scheduler pod, too

* Revert "Update k8s config to spin up a celerybeat scheduler pod, too"

This reverts commit 9d9c6e1caa97cfc9f46c4d375ba685f05964128b because
the better/correct approach is being done in a separate PR by @limed:
https://github.com/mdn/developer-portal/pull/422/

* Make how we get up the redis cache more robust, rather than assume a particular REDIS_URL configuration
+ remove default DB numbers from fallback Redis config.

If we get a REDIS_URL that specifies a particular DB, complain. Else build a cache-friendly Redis URL with a configurable DB number

Includes tests.

Shell check:

>>> from django.core import cache as cache_module
>>> cache_module.settings.CACHES
{'default': {'BACKEND': 'django_redis.cache.RedisCache', 'LOCATION': 'redis://redis:6379/1', 'OPTIONS': {'CLIENT_CLASS': 'django_redis.client.DefaultClient'}}}