Support filename hashing for S3 uploads
This changeset adds a custom get_upload_to method on `mozimages'` `MozImage` that takes the original filename and instead uses a slightly truncated SHA1 hash of it instead. This hashed filename is then respected in Renditions of the original image, too.

This change does NOT break existing files hosted on S3 but it may well break any image links on a local dev build that is not using S3.

Note that settings.AWS_S3_FILE_OVERWRITE is set to True to overcome django-storages default - see https://github.com/wagtail/wagtail/issues/1686