604: Enable HSTS by default
This changeset sets a default time of 1 hour for the Strict-Transport-Security
header, which we can upgrade to a year (31536000 seconds) once we're happy.

Note that the default behaviour for ALL non-local-development mode is to
have HSTS **ENABLED**.

Also, note that the SECURE_HSTS_INCLUDE_SUBDOMAINS setting is NOT set, so it
remains its default value of False. This is because Django warned (via
`manage.py check --deploy`) to "Only set this to True if you are certain
that all subdomains of your domain should be served exclusively via SSL." and
I don't know what else goes on `mozit.cloud` or `mozilla.com`.
Kuma doesn't appear to include subdomains in its config, either:
https://github.com/mdn/kuma/blob/master/kuma/settings/common.py#L1726

--

This changeset also removes a duplicated definition of SECURE_SSL_REDIRECT in settings.base