475: Disable the migration that drops the social-auth tables
...now that it has reached (and been run on) production.

If someone happens to still have the social-auth tables locally, that's not a great concern.
A fresh install will match the state of stage/prod.

Run locally to check that it can be applied forwards and backwards with no effect:
```
Operations to perform:
  Unapply all migrations: common
Running migrations:
  Rendering model states... DONE
  Unapplying common.0001_drop_social_auth_app_tables... OK
/app $ python manage.py migrate common
Operations to perform:
  Apply all migrations: common
Running migrations:
  Applying common.0001_drop_social_auth_app_tables... OK
```