718: Make it possible to see draft ExternalContent in the Admin
This changeset fixes an issue where draft pieces of ExternalContent were not visible in the Admin, which then can cause a server error when the user tries to add the same information again (with the same resulting slug, which causes a 500).

By ensuring the ExternalContent views show both live and draft states, we can avoid this.