864: Fix regression with generating static site
Somewhere between wagtail-bakery 0.3 and 0.4 (which was recently bumped as a dependency) a change took place that surfaces a shortcoming in our code: ExternalContent subclasses return None for .get_site() because they are not saved as children of Articles, Videos or Events, respectively.

The workaround for now is to just get the default Site for these Pages. A fuller fixup has been ticketed as #865.

Note that bumping django-bakery in this changeset didn't cause the regression - it was happening before that, and this change is just to bring it up to date, now that wagtail-bakery==0.4.0 allows a higher version of django-bakery.

I've also added a reminder to the developer notes about the downside of assuming we can always use the default Site if we lack a route to a Page's actual Site.

Manually checked a static build and seems fine.

Python tests pass; JS tests/linting checks pass