757: Support recording deployments in monitoring tools
This changeset repurposes code from mdn/kuma to add rollout tracking to developerportal.

It will ping New Relic and Speedcurve when a deployment has taken place.

Note that because a deployment doesn't automatically trigger a rebuild of the static site (but this does happen automatically every 30 mins) it's possible/likely that speedcurve will profile the previous static build. This isn't a massive problem, but is something to think about. We could instead trigger Speedcurve from the static build pipeline