825: Address SVG-related validation errors due to invalid hashes as ids
This changeset ensures all randomly generated hashes start with an `a` and are still 16 chars long.

There's no risk of collision because we're extending the original hash, rather than swapping out the
first (random) char for a fixed on.

This avoids validator.w3.org/nu reporting an error such as:
`Bad value 72644ef3cd99e4c20245b864b1ff1cb2-a for attribute id on element path: Not a valid XML 1.0 name.`