885: Ensure that if an Event has no Venue address specified, we don't show an empty location row
The use-case here is an Event in multiple places, where a single country cannot be configured. If there is no Venue address specified then prior to this changeset an empty UI element with a Location icon but no data would be rendered.

This changeset now makes that conditional: we only show the Location icon and row if there is data to populate it.

This changeset also includes showing the name of the venue in the location bar - it was previously only shown around the map box