906: Switch from Google Tag Manager to regular Google Analytics inclusion; add in Mozilla DNT support
This changeset moves away from using GTM and instead drops in GA code in the head of the document, as per mdn/kuma

At the same time it also uses the same Do Not Track lib that kuma uses, from @sneethling

Note that the DNT lib is slighty changed to pin a `Mozilla` object on the DOM's `window` so that it's within reachable scope within the page. The way Kuma integrates the same code is to inline it into the HTML (ie, render it as a template) but that behaviour isn't [currently?] suppored within Developer Portal.

As before, if there is no GOOGLE_ANALYTICS code present in settings, the relevant JS code block is not included. If it is present, the block is rendered. As such, only `settings.worker` will, by default, try to get GOOGLE_ANALYTICS from the env vars, so only the static site should render it out.

Example of included HTML/JS:

  <!-- Mozilla DNT Helper -->
  <script src="/static/js/head.4da485bfc790.js"></script>

  <!-- Only load GA if DNT is not enabled -->
  <script>
  if (window.Mozilla && !window.Mozilla.dntEnabled()) {
      (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
      (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
      m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
      })(window,document,'script','https://www.google-analytics.com/analytics.js','ga');

      ga('create', 'TestTestTest', 'mozilla.com');
      ga('set', 'anonymizeIp', true);
      (function() {
          // Remove analytics clutter from campaign URLs
          var win = window;
          var removeUtms = function(){
              var location = win.location;
              if (location.href.indexOf('utm') != -1 && win.history.replaceState) {
                  win.history.replaceState({}, '', location.pathname);
              }
          };
          ga('send', 'pageview', {'hitCallback': removeUtms});
      })();
  }
  </script>