886: Add template filter to move the cachebreaker string from inside the filename to a querystring
For the given URL, check if it has a cachebreaking hash. If it does,
convert it to be a URLthat uses the hash as a querystring cachebreaker
instead.

eg
    INPUT: 'https://example.com/static/foo.12313abc.jpg'
    OUTPUT: 'https://example.com/static/foo.jpg?h=12313abc'

Why? So that images automatically embedded in social media posts (via OG
tags) don't end up as broken links with a new release that changes the hash
in the filename -- by moving the hash to a querystring, we get the benefit
of cachebreaking but with a much softer edge that won't result in links
404ing even if the cachebreaking hash has changed. (ie, querystring
cachebreakers are more forgiving.)

Note that ManifestFileStorage (used by django-whitenoise in this project) makes
hash-named versions of the files, but also _does_ keep the original around,
so we are safe to assume that if static/img/foo.213123.jpg exists, then
static/img/foo.jpg will, too