Update requirements to fig regression when we moved to Wagtail 2.7
There were too many stale Wagtail-specific dependencies in the lower part of the requirements.txt file,
which meant that some things that should have been upgraded for wagtail 2.7 were not.

By removing the "subdeps" from requirements.txt wherever possible, we get a clean install.

The things that HAVE changed now are:
* modelcluster==5.0
* django-taggit==1.2.0
* python-dateutil==2.8.0  # down from 2.8.1 which helps with boto compat
* Willow==1.3

Note that beautifulsoup is pinned to the range that Wagtail needs, because readtime tries to get a newer, incompatible-with-Wagtail version