945: Update the filtering controls (left hand box) to NOT auto-submit when options are ticked
Auto filtering (when combined with network requests) was creating a grim UX, so we've switched to an explicit 'Refine results' button instead.

The decision was taken to use traditional form submission rather that needing JS to submit, which in turn required a minor refactor to how multiple values for querystrings were being sent and processed (ie from &topic=foo,bar to ?topic=foo&topic=bar).

This changeset also introduces whitelisting of filtering params - so we only listen to params we're expecting to receiv. (Just in case, really...)

Tests updated and manually checked, incl IE11 for the JS changes