994: Add a single source for the cutoff between past and future events, setting it to 'yesterday'
Prior to this changeset, the difference between future and past events is decided based
on datetime.datetime.now() which is likely to be UTC (given k8s), but regardless of what
specific timezone it is or is not, it still draws a single line "down the globe" that
means that, the further west of UTC you live, some events will 'disappear' from future
events before that day has completed.

The fix is simply to keep showing events as 'upcoming' until one calendar day *after*
their start date.