938: Ensure dynamic site renders out GA code
Prior to this changeset, we only included the Google Analytics code/JS in static pages, so
that we didn't end up having data beign sent from both the dynamic/CMS pages and also the
static ones.

This changeset updates the settings so that we get/use Google Analytics in all situations,
not just the static-site generation.