1017: Ensure tests are run using local file storage, not S3, plus update failing test
This came off the back of Travis CI failing up because it lacks S3 credentials and local-run tests (and Jenkins CI)was passing because S3 creds are available there.

While not ideal to test only on local FS, it is at least passing and using common Django filesystem APIs.

I've ticketed adding travis-only S3 creds to the Travis environment.