1017: Automatic content ingestion (#1100)
* WIP 1017: Lay in initial new app for handling ingestion

* add model and schema migration
* add data migration to populate initial configuration

* 1017: Add utils to handle parsing RSS and Atom feeds into lists of dicts, with tests

* 1017: Add support to ExternalArticle and ExternalVideo to create draft instances (saved) from data supplied from external feed

* 1017: Add support for ingesting content from feeds then converting into Page drafts, submitted for moderation

Includes tests

* 1017: Refactor: Change content target for one kind of import from ExternalVideo to Video

Also, support bootstrapping a 'Portal Bot' user to be the actor when importing data

Plus add a template with a simple meta-refresh redirect, so that ExternalArticles can be previewed

* 1017: Backfill some more tests

* 1017: Drop HTML meta-refresh way of redirecting to a preview...
because (since rebasing) we are overriding the Page.serve_preview() method on the ExternalContent base class

* 1017: Add support for enabling/disabling the automatic ingestion via env vars

* 1017: Ensure scheduled tasks for ingestion are run

Up till now, the scheduled tasks to ingest video or articles were not running, because there was no celerybeat configured for the `ingestion` app. The solution picked here was to register the tasks with the existing `taskqueue` app instead.
This is trading off coupling the ingestion and taskqueue apps in order to have just one celerybeat worker/pod, which makes life easier at the ops level.

We can revisit this if need be, no worries.

* 1017: Ensure tests are run using local file storage, not S3, plus update failing test

This came off the back of Travis CI failing up because it lacks S3 credentials and local-run tests (and Jenkins CI)was passing because S3 creds are available there.

While not ideal to test only on local FS, it is at least passing and using common Django filesystem APIs.

I've ticketed adding travis-only S3 creds to the Travis environment.

* 1017: Remove redundant check of model type

The removed code would not have been triggered because _get_model() was doing effectively the same check.

Merged the two so that _get_model() has a better output and dedicated tests

* 1017: Move the sending of notification emails out of the transaction

Rather than send one email each time a page is added from a feed, buffer the results and send the emails
later, outside of the transaction.

This is a bit cleaner and also leans more towards replacing invidiual emails with a single one, if we decide it's worth it

* 1017: Refactor generation of draft pages to be clearer, via factory funcs as arguments

Should be a lot easier to grok now.

Thanks for the suggestion @Gregoor!