1143: REFACTOR towards a JS-based showing of the notification based on presence of a cookie
This approach (while not final - see below) is because we can't rely on server-side rendering of the notification bar because the CDN will cache one version of each URI, which may or may not appropriately be showing the survey link.

The CDN can foward cookies, though, and just caches variants of each path according to the cookie that if finds accompanies the request/response. As such, we can use JS on the client side to check for the relevant Waffle cookie, then check its value and show the notification as appropriate.

NOTE: For now, until https://github.com/mozilla/protocol/issues/460 is resolved
it is not possible to render a hyperlink inside a NotificationBar that
is created by JS. That would be our ideal, because it means the CDN can
cache a version of the page with no NotificationBar, and we add it as
required. However, for now, we cannot do this AND have a link in it.

As a *temporary* workaround, we are rendering the notification with the link
in all HTML pages, but keeping it hidden by default, then revealing via JS
only to the appropriate users after checking for the relevant cookie