1143: Another attmept to stop the CDN inappropriately cacheing content re task completion survey
(If this works, tests and optimisations will come in the following commit)

In survey_tags.py, django-waffle will set a cookie as a result of checking
a flag. (It sets it either way, with a True or False value, depending on
whether we should show the survey to the user).

But there's a problem: if there's no cookie set on a request, Cloudfront wants
to cache this response as something it can use _any_ time it gets a
no-waffle-cookie request (and Cloudfront caches responses against a key
based on {request + cookies}, _not_ on {response + cookies}).

So, to deal with that, the _first_ time we are setting that a waffle cookie
for a user, we will tell CloudFront not to cache the cookies AND not to
cache the response.

This should mean that:
- if someone shows up with no waffle cookie set, we don't cache what
  they get back, but their next request will include a cookie that suits
  what they are supposed to get from Origin
- when we get a request WITH a waffle cookie of a certain value, if we have
  a match in the CDN cache, we CAN return it, and if we don't have a match
  the CDN will get it from Origin and cache it based on the request params,
  which is fine