1144: Remember when a user has dismissed the survey message for six weeks
This changeset adds code that makes use of the waffle cookie that determines whether a survey prompt should be shown or not. If its value is True, that means the survey is shown to the user. If it is False, the survey is not shown, but the cookie persists until its expiry date.

Now, when a user clicks the "X" to close/dismiss the notification, this rewrites the value stored in the cookie, to be False, not True. It also has to set a new expiry time, else it'll only exist for the life of the browser session, so we set it to six weeks from now, meaning the user will have a cookie that says "Don't ask me about a survey" for six weeks, after which time it will expire and the user _might_ then (depending on the percentage settings of the survey waffle flag on the server) see a subsequent request to complete a survey - but it will at least have been six weeks.

Note that the decision was taken to ONLY tweak the cookie when the notification is dismissed, and not when the user clicks on the link to view the survey. This seemed to be more sensible in terms of giving users a chance to lose the link/tab and choose to view it again, while also allowing them to expressly dismiss the notificaition.