1207: Add support for styling specified links as buttons in Draftail editor
(Note: a big hat-tip goes to @kalobtaulien for some his assistance with this, making it a much smaller change than writing an entirely new Draftail Entity.)

This changeset supports a way of marking certain links in a rich-text area as buttons, so that they are rendered as such in the published page.

(This should allow us to make pages that feature lots of buttons without the page becoming too heavy/complex, which appears to be a factor in editing hanging after a certain tipping point.)

It acheives this by adding a new button to the Draftail editor which adds a new block (a div, lightly styled with a dashed grey border) inside which one can add hyperlinks using rich-text editor's standard link button. Any links inside this block will be rendered as buttons in the published (and previewed) page, and are lightly styled to look similar within the rich-text editor itself.