1143 + 1144: Task completion survey (core implementation) (#1150)
* 1143: Task-completion survey work: add dismissable notification bar with dummy link for now

Uses the notification bar from Mozilla Protocol.

Checked on desktop and mobile viewports. I've kept the CTA text brief, to aid with reflow issues, but there are still some non-standard viewport widths where the notification text flows badly - though behaviour at our standard breakpoints is solid.

* 1143: Support obtaining the task-completion survey URL from configuration

Can be set in an env var, or overridden in production.py

* 1443: Open survey link in a new tab/window, in a safe way

Using "nofollow" because we're linking off to a third party site for the survey
Using "noopener" to avoid exploitation of the window.opener API and the page performance hit

* 1143: Only show the survey link bar if there is a survey link configured

* 1143: Add django-waffle, and use to show survey to 5% of visitors

Includes data migration to set up that flag

* 1143: Add tests for current use of django-waffle and also for absence of configured survey URL

* 1143: Add support for env var configuration of survey URL in k8s

Including adding a dummy value for the dev env, and adding commented-out placeholders for stage and prod

Also includes a behaviour tweak to cope with the URL being set to "undefined" via a Makefile default

* 1143: Update local settings template to handle a gotcha related to secure cookies and local non-HTTPS development

* 1143: REFACTOR towards a JS-based showing of the notification based on presence of a cookie

This approach (while not final - see below) is because we can't rely on server-side rendering of the notification bar because the CDN will cache one version of each URI, which may or may not appropriately be showing the survey link.

The CDN can foward cookies, though, and just caches variants of each path according to the cookie that if finds accompanies the request/response. As such, we can use JS on the client side to check for the relevant Waffle cookie, then check its value and show the notification as appropriate.

NOTE: For now, until https://github.com/mozilla/protocol/issues/460 is resolved
it is not possible to render a hyperlink inside a NotificationBar that
is created by JS. That would be our ideal, because it means the CDN can
cache a version of the page with no NotificationBar, and we add it as
required. However, for now, we cannot do this AND have a link in it.

As a *temporary* workaround, we are rendering the notification with the link
in all HTML pages, but keeping it hidden by default, then revealing via JS
only to the appropriate users after checking for the relevant cookie

* 1143: Further refactoring to alwasy render the hidden notification, to reduce the size of the tree of option while debugging CDN issues

* 1143: Update tests to match changes in previous commit

* 1143: Move Waffle flag name into constants, ready to re-use elsewhere

* 1143: Another attmept to stop the CDN inappropriately cacheing content re task completion survey

(If this works, tests and optimisations will come in the following commit)

In survey_tags.py, django-waffle will set a cookie as a result of checking
a flag. (It sets it either way, with a True or False value, depending on
whether we should show the survey to the user).

But there's a problem: if there's no cookie set on a request, Cloudfront wants
to cache this response as something it can use _any_ time it gets a
no-waffle-cookie request (and Cloudfront caches responses against a key
based on {request + cookies}, _not_ on {response + cookies}).

So, to deal with that, the _first_ time we are setting that a waffle cookie
for a user, we will tell CloudFront not to cache the cookies AND not to
cache the response.

This should mean that:
- if someone shows up with no waffle cookie set, we don't cache what
  they get back, but their next request will include a cookie that suits
  what they are supposed to get from Origin
- when we get a request WITH a waffle cookie of a certain value, if we have
  a match in the CDN cache, we CAN return it, and if we don't have a match
  the CDN will get it from Origin and cache it based on the request params,
  which is fine

* 1143: Refactor CDN-cache-skipping middleware (re waffle flag for survey) and add tests

* 1143: Only trigger the waffle-flag check (and therefore cookie setting) if we have a survey URL defined at all

* 1143: Update the note set for the Waffle Flag, to explain to leave most fields in their default state

* 1143: IE11 FIX -> use compliant for loop syntax, rather than ignore it.

The waring is there for a reason! The buble config says `transforms: { forOf: false }`

* 1143: Add Object.remove polyfill to make the Mozilla Protocol NotificationBar work in IE11

ref: https://github.com/mozilla/protocol/issues/535

* 1143: Drop use of django-waffle after all

The downside having to skip the CDN for the first one or two requests from a new user for every page is just too much to weather, so we're going with a simpler approach to showing the task-completion survey.

This commit pulls out django-waffle and related code, leaving the survey-prompt code non-functional - normal service will be restored in the next commit

* 1143: Refactor task-completion-survey logic to make it entirely client-side

* Decide whether to show the prompt based on a percentage passed from Django to JS
* Remember this decision in a cookie with a six-week lifespan
* If the cookie is set, respect it - don't re-roll the dice
* When a user dismisses the notification, update the cookie to be 'False'/don't show, and continue to remember this for six weeks

To come in subsequent commits:
* k8s config
* unit tests for the JS side
* test tweaks for the Django side

* 1143: Add new percentage-of-people-to-show-survey-to env var to k9s config

* 1143: Stop using homespun code for cookie management - switch to js-cookie package (MIT)

Also slightly separate out the task-completion code

* 1143: Add some light tests for the JS related to the task survey prompt

Not exhaustive, but some regression protection.

* 1143: Ensure the hover colour of the CTA link in a notification remains visible, rather than becoming blue on blue

* 1143: Clean up logic for getting a boolean weighted by percentage

Thanks @Gregoor

* 1143: Fix misnomer spotted by @Gregoor - thanks!

* 1143: Improve logic in deciding whether to show a survey or not, based on cookie

Hat-tip to @Gregoor

* 1143: Add more test coverage for the task completion prompt being revealed